% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hellomojo}
\alias{hellomojo}
\title{Hello from Mojo
Call the native 'hello' function from the Mojo shared library}
\usage{
hellomojo()
}
\description{
Hello from Mojo
Call the native 'hello' function from the Mojo shared library
}
