% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hellomojo_add}
\alias{hellomojo_add}
\title{Add two numbers using the native 'add' function from the Mojo shared library}
\usage{
hellomojo_add(a, b)
}
\arguments{
\item{a}{A numeric value}

\item{b}{A numeric value}
}
\value{
The sum of a and b
}
\description{
Add two numbers using the native 'add' function from the Mojo shared library
}
