% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hellomojo_convolve}
\alias{hellomojo_convolve}
\title{1D convolution using the native Mojo function}
\usage{
hellomojo_convolve(signal, kernel)
}
\arguments{
\item{signal}{Numeric vector (signal)}

\item{kernel}{Numeric vector (kernel)}
}
\value{
Numeric vector (convolution result)
}
\description{
1D convolution using the native Mojo function
}
