% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojo_build.R
\name{mojo_build_package}
\alias{mojo_build_package}
\title{Build Mojo library for this package}
\usage{
mojo_build_package(
  venv = ".venv/mojo",
  source = "inst/mojo/hellomojo/hellomojo.mojo"
)
}
\arguments{
\item{venv}{Path to virtual environment with Mojo}

\item{source}{Path to Mojo source file}
}
\description{
Compiles the Mojo source and installs it to inst/libs
}
