% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojo_build.R
\name{mojo_info}
\alias{mojo_info}
\title{Get information about Mojo installation}
\usage{
mojo_info(venv = NULL)
}
\arguments{
\item{venv}{Path to virtual environment}
}
\description{
Get information about Mojo installation
}
