% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojo_build.R
\name{mojo_install}
\alias{mojo_install}
\title{Create a Python virtual environment and install Mojo}
\usage{
mojo_install(venv = ".venv/mojo", nightly = TRUE, python = "python3")
}
\arguments{
\item{venv}{Path where to create the virtual environment}

\item{nightly}{If TRUE, install nightly build. If FALSE, install stable.}

\item{python}{Python executable to use. Default is "python3".}
}
\description{
Create a Python virtual environment and install Mojo
}
